refac(Utils): Create static methods to reuse different methods in jobs
