--- Assets/Scenes/Grid2D/BakedGrid.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e6e9e34d0e09c4343a45e23639d2d3e0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
