--- Assets/Scripts/Agents/AgentHighUpdateJob.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ed0cf00d4f28c604f8d3bf14df6c3239
