--- Assets/Scripts/Agents/AgentLowUpdateJob.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ca5105e4b00ff4e4c86b7365a84af38b
