+++ Assets/Scripts/Utilities/Grid2DUtils.cs
@@ -1,0 +1,9 @@
+using NavigationGraph;
+
+public static class Grid2DUtils
+{
+    public static Cell GetCurrentCell()
+    {
+        return new Cell(); // Placeholder implementation
+    }
+}
