+++ Assets/Scripts/Utilities/MathUtils.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3763035d90c1d994c8c1166261528d06
